# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transport/tenants:
    get:
      operationId: listTenants
      tags: [transport-tenants]
      summary: List all tenant companies
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [transport-tenants]
      summary: Create a new tenant company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantBody"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /transport/tenants/{tenantId}:
    get:
      operationId: getTenant
      tags: [transport-tenants]
      summary: Get a tenant by ID
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
    patch:
      operationId: updateTenant
      tags: [transport-tenants]
      summary: Update a tenant company
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantBody"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

  # ─── USER MANAGEMENT ──────────────────────────────────────────────────────
  /transport/users:
    get:
      operationId: listTransportUsers
      tags: [transport-users]
      summary: List users (filtered by tenant for non-superadmin)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransportUser"
    post:
      operationId: createTransportUser
      tags: [transport-users]
      summary: Create a new transport user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransportUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransportUser"
  /transport/users/{userId}:
    patch:
      operationId: updateTransportUser
      tags: [transport-users]
      summary: Update a transport user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransportUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransportUser"

  # ─── DISPATCHES ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TransportUser:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: ["integer", "null"]
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [superadmin, admin, controlador]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - tenantId
        - email
        - name
        - role
        - isActive
        - createdAt
        - updatedAt
    CreateTransportUserBody:
      type: object
      properties:
        tenantId:
          type: ["integer", "null"]
        email:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, controlador]
      required:
        - email
        - password
        - name
        - role
    UpdateTransportUserBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
    Tenant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nit:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        address:
          type: string
        city:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - nit
        - contactName
        - contactEmail
        - contactPhone
        - address
        - city
        - isActive
        - createdAt
        - updatedAt
    CreateTenantBody:
      type: object
      properties:
        name:
          type: string
        nit:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        address:
          type: string
        city:
          type: string
      required:
        - name
    UpdateTenantBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        nit:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
...