# lib/api-zod/src/generated/types/tenant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Tenant {
  id: number;
  name: string;
  nit: string;
  contactName: string;
  contactEmail: string;
  contactPhone: string;
  address: string;
  city: string;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}